extract logo into business component
