+++ packages/generic/client/src/components/business/index.ts
@@ -1,3 +1,4 @@
 import { CategoryIcon } from './CategoryIcon';
+import { Logo } from './Logo';
 
-export { CategoryIcon };
+export { CategoryIcon, Logo };
