+++ packages/generic/client/src/components/generic/materialUI/SvgImage.tsx
@@ -1,0 +1,8 @@
+import React, { ReactNode, CSSProperties } from 'react';
+import { SvgIcon } from '@material-ui/core';
+
+export { SvgImage };
+
+function SvgImage(props: { children: ReactNode; style: CSSProperties }) {
+  return <SvgIcon style={props.style}>{props.children}</SvgIcon>;
+}
