--- packages/generic/client/src/components/generic/materialUI/icons/index.ts
@@ -1,5 +1,0 @@
-import { Icon, iconNameType } from './Icon';
-import { LogoIcon } from './LogoIcon';
-
-export { Icon, LogoIcon };
-export type { iconNameType };
