+++ packages/generic/client/src/components/generic/materialUI/index.ts
@@ -1,26 +1,28 @@
 import { Button, SwitchButton } from './buttons';
+import { Accordion } from './Accordion';
 import { Checkbox } from './Checkbox';
 import { Dropdown } from './Dropdown';
+import { Icon } from './Icon';
 import { LayoutGrid } from './LayoutGrid';
 import { MenuBar } from './MenuBar';
-import { Accordion } from './Accordion';
-import { Icon } from './icons';
+import { SvgImage } from './SvgImage';
 import { Text } from './Text';
 import { TextInput } from './TextInput';
 import { TextInputLarge } from './TextInputLarge';
 import { TooltipMenu } from './TooltipMenu';
 
 export {
+  Accordion,
   Button,
-  SwitchButton,
   Checkbox,
   Dropdown,
+  Icon,
   LayoutGrid,
   MenuBar,
+  SvgImage,
+  SwitchButton,
+  Text,
   TextInput,
   TextInputLarge,
-  Accordion,
-  Icon,
-  Text,
   TooltipMenu,
 };
